Moved AuthContext into components
