+++ frontend/src/styles/Settings.css
@@ -8,4 +8,6 @@
 .content-settings {
   display: grid;
   flex: 1;
+  grid-template-columns: repeat(4, 1fr);
 }
+
